# ui/openapi/bizigo-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bizigo.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/alerts/triggers/{triggerId}/close": {
      "post": {
        "tags": [
          "alerts"
        ],
        "operationId": "CloseAlertTrigger",
        "parameters": [
          {
            "name": "triggerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloseTriggerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloseTriggerRequest": {
        "type": "object",
        "properties": {
          "verdict": {
            "type": "string"
          },
          "contradicting_evidence": {
            "type": "string"
          },
          "actual_root_cause": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "correct_finding_rank": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "rank_asked": {
            "type": "boolean"
          },
          "missing_evidence_kinds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "CloseTriggerResponse": {
        "required": [
          "trigger_id",
          "closed_at",
          "bundle_generated",
          "review_id",
          "owner_group"
        ],
        "type": "object",
        "properties": {
          "trigger_id": {
            "type": "string",
            "format": "uuid"
          },
          "closed_at": {
            "type": "string",
            "format": "date-time"
          },
          "bundle_generated": {
            "type": "boolean"
          },
          "review_id": {
            "type": "string",
            "format": "uuid"
          },
          "owner_group": {
            "type": "string"
          },
          "schema_version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 3
          },
          "missing_evidence_kinds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "missing_evidence_asked": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "hint": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}